Switch `ktime_get_ns()` to tried-and-true `ktime_get()`.
Seems we're getting a segfault during our `ktime_get_ns()` call in the
previous commit, so we revert to a profiling method we've seen before.
If there are errors after this build, then `ktime_get()` must segfault
during early boot. If there are no errors, then the ns version was the
problem.